js updated .gitignore
