--- PLAN.md
@@ -1,3 +1,0 @@
-hey guys
-
-Google Doc Link to Project Architecture Picture: https://docs.google.com/document/d/1bHNQ3UmWAYq2u3JhgFjrHw29OtLaLDJNvfh6tMkLdOk/edit?usp=sharing
