--- sampleOutput/groups.csv
@@ -1,24 +1,0 @@
-6987,533,510
-5564,185,453
-5187,586,195
-5115,371,244
-7,617,229
-7,377,280
-6,592,233
-5,611,231
-5,573,235
-5,373,286
-4,599,232
-4,582,234
-2,564,236
-2,507,552
-2,501,556
-2,497,558
-2,196,487
-1,551,550
-1,546,551
-1,371,281
-1,368,282
-1,228,470
-1,167,494
-1,152,498
