--- src/main/java/io/github/oakes777/salamander/VideoFrame.java
@@ -1,8 +1,0 @@
-package io.github.oakes777.salamander;
-
-import java.awt.image.BufferedImage;
-
-/**
- * Simple wrapper to hold a video frame and its timestamp (in seconds).
- */
-public record VideoFrame(BufferedImage image, double timestampSeconds) {}
